refactor: decompose timer.tsx into hook and components
Extract useTimerEngine hook, TimerRing, ConfigPanel, and TimerControls
components to reduce FTA complexity. Timer.tsx shrinks from 706 to 225
lines while preserving all existing behavior.
